% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{check_duplicates.smi}
\alias{check_duplicates.smi}
\title{check that the smiles and compound columns are unique}
\usage{
check_duplicates.smi(smi, smiles = "smiles", compound = "compound")
}
\description{
check that the smiles and compound columns are unique
}

