% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{clean_duplicates.smi}
\alias{clean_duplicates.smi}
\title{Remove duplicate entries in a smiles dataset}
\usage{
clean_duplicates.smi(smi, smiles = "smiles", compound = "compound")
}
\description{
Remove duplicate entries in a smiles dataset
}
