% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{cleanup.config_files}
\alias{cleanup.config_files}
\title{Cleanup SEA config files}
\usage{
cleanup.config_files(dir = getwd(), fitcore = TRUE, fpcore = TRUE,
  seacore = TRUE, sea_interface = TRUE)
}
\description{
Cleanup SEA config files
}

