% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sea_scores.R
\name{collect_compute_sea_network}
\alias{collect_compute_sea_network}
\title{Collect target vs. target scores computed with submit_compute_sea_network
Add Q-values using the fdrtool package}
\usage{
collect_compute_sea_network(run_base, verbose)
}
\arguments{
\item{run_base}{returned by submit_compute_sea_network(...)}
}
\description{
Collect target vs. target scores computed with submit_compute_sea_network
Add Q-values using the fdrtool package
}
