% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create.fp}
\alias{create.fp}
\title{create a fingerprint file for the given file or data frame with columns [compound, smiles]
return based on the read parameter}
\usage{
create.fp(smi, compound = "compound", smiles = "smiles",
  output_fname = NULL, fp_format = c("sea_native", "bitstring"),
  fp_type = c("rdkit_ecfp", "rdkit_path", "rdkit_ecfp_mask"),
  standardize = T, verbose = F)
}
\description{
create a fingerprint file for the given file or data frame with columns [compound, smiles]
return based on the read parameter
}

