% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{event_probs}
\alias{event_probs}
\title{Extract probabilities from data frame and risk set}
\usage{
event_probs(data, identities, probs = NULL, log = FALSE)
}
\arguments{
\item{data}{data frame}

\item{identities}{identities from risk set}

\item{probs}{external vector of probabilities (optional)}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
a vector of conditional probabilities (or log-probabilities)
}
\description{
Extract probabilities from data frame and risk set
}
