% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{generate_images}
\alias{generate_images}
\title{Generate svg images for each compound}
\usage{
generate_images(smi, smiles = "smiles", compound = "compound",
  verbose = F, batch_size = 1000)
}
\description{
Generate svg images for each compound
}
