% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{genesets_mirs}
\alias{genesets_mirs}
\title{10,786 annotated gene sets}
\format{A data frame with 10786 rows and 4 variables:
\itemize{
  \item collection: origin of the gene set (MSigDB: H_hallmark, C1_position -
  C7_immunologic collections, BTM or TISSUES)
  \item subcollection: collections are sometimes further subdivided (e.g.
  C2_canonical: CP_BIOCARTA, CP_KEGG, CP_PID, CP_REACTOME, etc.)
  \item geneset: name of the gene set in the collection
  \item members: miRNA members (using naming convention used by mirBase; described:
  Victor Ambros, Bonnie Bartel, David P. Bartel, Christopher B. Burge, James
  C. Carrington, Xuemei Chen, Gideon Dreyfuss, Sean R. Eddy, Sam
  Griffiths-Jones, Mhairi Marshall, Marjori Matzke, Gary Ruvkun, and Thomas
  Tuschl. A uniform system for microRNA annotation. RNA 2003 9(3):277-279.)
}}
\usage{
genesets_mirs
}
\description{
A dataset containing nearly 11,000 annotated gene sets from the following
sources:
}
\details{
\itemize{
  \item MSigDB: Subramanian, Aravind, et al. "Gene set enrichment
analysis: a knowledge-based approach for interpreting genome-wide expression
profiles." Proceedings of the National Academy of Sciences of the United
States of America 102.43 (2005): 15545-15550.
  \item Blood Transciptome
Modules (BTM): Li, Shuzhao, et al. "Molecular signatures of antibody
responses derived from a systems biology study of five human vaccines."
Nature immunology 15.2 (2014): 195-204
  \item Tissue Enrichment: Benita, Yair,
et al. "Gene enrichment profiles reveal T-cell development, differentiation,
and lineage-specific transcription factors including ZBTB25 as a novel NF-AT
repressor." Blood 115.26 (2010): 5376-5384.
}
Gene sets were mapped to equivalent validated targetting miRNAs using the
approach proposed in:
...

Variables are as follows:
}
\keyword{datasets}

