% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_identities}
\alias{get_identities}
\title{Extract Identities from risk set object}
\usage{
get_identities(rs)
}
\arguments{
\item{rs}{a risk set object}
}
\value{
a character vector
}
\description{
Extract Identities from risk set object
}
