% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{H}
\alias{H}
\title{Shannon Index of Diversity}
\usage{
H(data, identities)
}
\arguments{
\item{data}{data frame with .id column}

\item{identities}{vector of identity names}
}
\value{
a vector of diversity indicators (entropies)
}
\description{
Shannon Index of Diversity
}
