% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{hit.report}
\alias{hit.report}
\title{Generate a hit report using the BioChemPantry
scores is a data.frame with columns target1, target2 as uniprot_entries}
\usage{
hit.report(hits)
}
\description{
Generate a hit report using the BioChemPantry
scores is a data.frame with columns target1, target2 as uniprot_entries
}
