% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Numerically stable computation of log sums of exponentiated values.}
\source{
Taken from Richard McElreath's \code{rethinking} package
}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
a scalar
}
\description{
This helps to avoid rounding errors that occur when working with direct probabilities.
}
