% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{pack.library}
\alias{pack.library}
\title{Pack a SEA library}
\usage{
pack.library(molecules, targets, library_fname = NULL,
  fingerprint_type = "rdkit_ecfp4", name = NULL, config_files = NULL,
  verbose = F)
}
\description{
Pack a SEA library
}

