% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{plot_fits.library}
\alias{plot_fits.library}
\title{Fit a library and generate plots}
\usage{
plot_fits.library(library_fname, diagnostics_fname = library_fname,
  verbose = F)
}
\description{
Fit a library and generate plots
}

