% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mtelog.R
\name{read_mtelog}
\alias{read_mtelog}
\title{Read and tidy data from Applanix, CTD, BB3, SeaOWL, SUNA}
\usage{
read_mtelog(LogFile)
}
\description{
The read function take as input MainLog, a first parsing of the txt file of
the data logger
}
