% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{risk_set}
\alias{risk_set}
\title{Risk Set}
\usage{
risk_set(size, ..., sep = "|")
}
\arguments{
\item{size}{a numeric vector}

\item{...}{vectors of identities}

\item{sep}{a string}
}
\value{
a risk set data frame
}
\description{
Risk Set
}
