% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{risk_set_size}
\alias{risk_set_size}
\title{Risk Set Sizes}
\usage{
risk_set_size(size, ...)
}
\arguments{
\item{size}{a numeric vector}

\item{...}{vectors of identities}
}
\value{
a vector of sizes
}
\description{
Risk Set Sizes
}
