% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{run.sea}
\alias{run.sea}
\title{run SEA}
\usage{
run.sea(background_fname, ref_set, ref_smi, query_set, query_smi,
  run_tag = NULL, output_base = NULL, similarity_measure = "tanimoto",
  tversky_alpha = 0.5, tversky_beta = 0.5)
}
\description{
inputs:
  background_fname: filename of .fit file
  ref_set: compound sets for reference targets
    path to .set file or
    data.frame with columns [target, name, compound]
  ref_smi: compounds for reference targets
    path to .smi file or
    data.frame with columns [compound, smiles]
  query_set: compound sets for query targets
    path to .set file or
    data.frame with columns [target, name, compound]
  query_smi: compounds for query targets
    path to .smi file or
    data.frame with columns [compound, smiles]
  run_tag: a tag given help identify the sea analysis in the SEA viewer
    default NULL
  output_base: path and file prefix where the files should be written
    default to a temporary file
}
\details{
Output:
  data.frame with columns [target1, target2, EValue, MaxTC] for SEA associations
export
}

