% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sea_viewer_locator}
\alias{sea_viewer_locator}
\title{Locate the SEA viewer
build redirect urls into the sea viewer a list of target pairs}
\usage{
sea_viewer_locator(query_targets, query_activity_thresholds, reference_targets,
  reference_activity_thresholds,
  sea_locator_url = "http://sea16.ucsf.bkslab.org/custom/search/target_vs_target")
}
\description{
sea_locator_url: url to the locate_association.php script
   you may need to copy this file into the sea viewer base directory.
 sea_analsis_id: each sea run is assigned an analysis id
 query_targets, reference_targets: vectors of target codes
 flip: if true put the query target downstairs
 return:
    a character vector of urls into the SEA viewer
}
