% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sear-package.R
\docType{package}
\name{sear-package}
\alias{sear-package}
\alias{_PACKAGE}
\alias{testduplicate-package}
\title{sear: Simple (effin') enrichment analysis in R}
\description{
A package for simple enrichment analysis using: 1.MSigDB: Subramanian, Aravind, et al. "Gene set enrichment analysis: a knowledge-based approach for interpreting genome-wide expression profiles." Proceedings of the National Academy of Sciences of the United States of America 102.43 (2005): 15545-15550. 2.Blood Transciptome Modules (BTM): Li, Shuzhao, et al. "Molecular signatures of antibody responses derived from a systems biology study of five human vaccines." Nature immunology 15.2 (2014): 195-204 3.Tissue Enrichment: Benita, Yair, et al. "Gene enrichment profiles reveal T-cell development, differentiation, and lineage-specific transcription factors including ZBTB25 as a novel NF-AT repressor." Blood 115.26 (2010): 5376-5384.
}
\author{
\strong{Maintainer}: Casey Shannon \email{also.casey@gmail.com}

}
\keyword{internal}
