% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sear.r
\name{sear}
\alias{sear}
\title{Simple enrichment analysis in R using the MSigDB collections.}
\usage{
sear(genes)
}
\arguments{
\item{genes}{A character vector of HUGO gene naming commitee (HGNC) gene
symbols.}

\item{genesets}{A data.frame of annotated gene sets: defaults to a prepared
aggregate of the collections described above.}
}
\description{
Carry out enrichment analysis against some or all of the MSigDB collections
(1), Blood Transcriptome geness (2) and Tissue Enrichment Sets (3).
}
\details{
A list of genes is compared to each annotated gene set in turn by performing
a hypergeometric test of the overlap. The size of the input gene list, gene
set, intersection and resulting p-value are returned. P-values can be
adjusted over each collection (optimistic) or globally (pessimistic), using
any of the methods available to `p/adjust()`.
}
\examples{
sear(c("TTLL4", "CTSK", "NNMT", "TGFBR2", "MAGED2", "ASB13", "CCDC80", "APBB2", "RABEP1", "FBP1"))
}

