% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{set_fit.library}
\alias{set_fit.library}
\title{Set the library fit}
\usage{
set_fit.library(library_fname, threshold = 0.28,
  diagnostics_fname = library_fname, verbose = F)
}
\arguments{
\item{library_fname}{The name of the sea library}

\item{threshold}{the Tc Threshold to use
diagnostics_fname the file <diagnostics_fname>.dist should exist it is generated by plot_fits.library}
}
\description{
Set the library fit
}

