% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_collapse}
\alias{sort_collapse}
\title{Sort + Collapse}
\usage{
sort_collapse(..., sep = "|")
}
\arguments{
\item{...}{a set of variables}

\item{sep}{defaults to "|"}
}
\value{
a new id variable
}
\description{
This functions takes a set of character variables and collapses them into one
unique identifier after appropriate sorting. The resulting id will correspond
to the one produced by \code{risk_set()}
}
