% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sea_scores.R
\name{submit_compute_sea_network}
\alias{submit_compute_sea_network}
\title{Compute target vs. target scores for all targets in the library
This sets up the sge cluster job that is then submitted from the command line:}
\usage{
submit_compute_sea_network(library_fname, library_id, fp_type, runs_dir = ".",
  config_files = NULL, target_range = "all", verbose = TRUE,
  dry_run = FALSE)
}
\arguments{
\item{library_fname}{.sea file
generated with SEAR::pack.library(...)
and then the background fit with plot_fits.library(...) and then set_fit.library(...)}
}
\value{
run_base
}
\description{
run_base <- <runs_dir>/<library_id>_vs_<library_id>
}
\details{
cd <run_base>/logs
  qsub -t 1-<n_targets> <path/to/sea-wrapper.sh> <run_base> <library_fname> <fp_type>

then run SEAR::collect_compute_sea_network(...) to generate scores file
}
