% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{tc_matrix}
\alias{tc_matrix}
\title{Compute a sparse TC matrix between query and reference compounds with values above the Tc cutoff
ref_fp: Either
   a) path to a fingerprints file as generated with create.fp
   b) data.frame with given column <ref_compound> and <ref_smiles>.
query_fp: Either
   a) path to a fingerprints file as generated with create.fp
   b) data.frame with given column <ref_compound> and <ref_smiles>.
Requires seaware-academic to be loaded
returns data.frame with columns <compound>, <MaxTC> for all query compounds with Tc >= cuttoff  of a reference compound}
\usage{
tc_matrix(ref_fp, query_fp, cutoff = 0, output_fname = NULL,
  ref_compound = "compound", ref_smiles = "smiles",
  query_compound = "compound", query_smiles = "smiles", verbose = F, ...)
}
\description{
Compute a sparse TC matrix between query and reference compounds with values above the Tc cutoff
ref_fp: Either
   a) path to a fingerprints file as generated with create.fp
   b) data.frame with given column <ref_compound> and <ref_smiles>.
query_fp: Either
   a) path to a fingerprints file as generated with create.fp
   b) data.frame with given column <ref_compound> and <ref_smiles>.
Requires seaware-academic to be loaded
returns data.frame with columns <compound>, <MaxTC> for all query compounds with Tc >= cuttoff  of a reference compound
}
