% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{teams}
\alias{teams}
\title{Teams}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 816 rows and 14 columns.
}
\source{
Martin
}
\usage{
teams
}
\description{
A subset of data from the ...
}
\keyword{datasets}
