% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot_entry_web_lookup}
\alias{uniprot_entry_web_lookup}
\title{See http://www.uniprot.org/help/uniprotkb_column_names for available columns}
\usage{
uniprot_entry_web_lookup(uniprot_entries, columns, user_agent_arg,
  verbose = F)
}
\description{
See http://www.uniprot.org/help/uniprotkb_column_names for available columns
}

