% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{unpack.library}
\alias{unpack.library}
\title{Unpack a SEA library
returns: list(molecules, targets) where
  molecules is a data.frame with columns: compound, smiles, fingerprint
  targets is a data.frame with columns: target, name, affinity, compound, description}
\usage{
unpack.library(fname, fingerprint_format = "sea_native",
  molecules_fname = NULL, targets_fname = NULL, unpack_compounds = TRUE,
  verbose = FALSE)
}
\description{
Note: the molecules and targets tables are joined on the compound in the molecules and targets is the join column
}

