% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.config_files}
\alias{write.config_files}
\title{Write SEA config files
Config files should be a list <config_fname>=<config_file_contents>
for one of the sea config files.
these are written to the current working directory so they can be parsed
so as to take presidence over other specified configuration options}
\usage{
write.config_files(dir = getwd(), fitcore = NULL, fpcore = NULL,
  seacore = NULL, sea_interface = NULL)
}
\description{
Write SEA config files
Config files should be a list <config_fname>=<config_file_contents>
for one of the sea config files.
these are written to the current working directory so they can be parsed
so as to take presidence over other specified configuration options
}

