% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.set}
\alias{write.set}
\title{Write a set file
INPUT:
  set: data.frame with columns [target, name, affinity, description, compound]
       if affinity and description are not present, assume it is the old style
       format and don't write them out.}
\usage{
write.set(set, fname, target = "target", name = "name",
  compound = "compound", affinity = "affinity",
  description = "description", seaware_format = T)
}
\description{
warning note the order of the columns in the input idfferes in the order of the output
  set_fname: .set file name
OUTPUT:
  creates <set_fname>
}

