% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.singleton.set}
\alias{write.singleton.set}
\title{Write a singleton sets file}
\usage{
write.singleton.set(smi, fname, target = "compound", name = "compound",
  compound = "compound")
}
\description{
Write a singleton sets file
}
