% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.smi}
\alias{write.smi}
\title{input data.frame with columns [smiles, compound]
write .smi file to given fname}
\usage{
write.smi(smi, fname, smiles = "smiles", compound = "compound",
  seaware_format = T, verbose = F)
}
\description{
filter unique compound identifiers
}

