% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.r
\name{clear}
\alias{clear}
\title{Clear}
\usage{
clear(leading_edge, cutoff = 0.25, trim = FALSE)
}
\arguments{
\item{leading_edge}{The output of running sear on an input list of mRNAs or
miRNA symbols.}

\item{cutoff}{A numeric value between 0 and 1. The minimum value of the
jaccard coefficient to show links between genesets.}

\item{trim}{A logical value. Whether to perform trimming of non-connected
nodes.}
}
\description{
Clear
}
\examples{

library(dplyr)
data('genesets')
input <- genesets \%>\%
  filter(collection == 'tissues') \%>\%
  nth(4) \%>\%
  unlist() \%>\%
  unique() \%>\%
  sample(100)
output <- sear(input, type = 'mrna') \%>\%
  arrange(fdr) \%>\%
  slice(1:100)
clear(output, cutoff = 0.25, trim = TRUE)
}

