% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create.fp}
\alias{create.fp}
\title{create a fingerprint file for the given file or data frame with columns [compound, smiles]
return based on the read parameter
  read == 'none' -> return NULL
  read == 'fname' -> return the filename of the generated fingerprint file
  read == 'compound' -> return just the list of compounds in the fingerprint file
  read == 'full' -> read in the full fingerprint file}
\usage{
create.fp(smi, read = c("none", "fname", "compound", "full"),
  compound = "compound", smiles = "smiles", output_fname = NULL)
}
\description{
create a fingerprint file for the given file or data frame with columns [compound, smiles]
return based on the read parameter
  read == 'none' -> return NULL
  read == 'fname' -> return the filename of the generated fingerprint file
  read == 'compound' -> return just the list of compounds in the fingerprint file
  read == 'full' -> read in the full fingerprint file
}

