% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time.R
\name{data_synthesis}
\alias{data_synthesis}
\title{data_synthesis}
\usage{
data_synthesis(x, y, tol = 3)
}
\arguments{
\item{x}{time vector of the 'Main' log, usaly GNSS data).}

\item{y}{time vector of the instrument for which to create the synthesis}

\item{tol}{allowed time difference (+-)}
}
\value{
A logical vector (T/F),
}
\description{
Create a data synthesis from two time vectors. Give information on which data have been acquired when.
}
