% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{genesets_links}
\alias{genesets_links}
\title{Adjacency of 10,786 annotated gene sets as a dyadic table}
\format{A data frame with 22,081,622 rows and 3 variables:
\itemize{
  \item source: ...
  \item target: ...
  \item jaccard: ...
}}
\usage{
genesets_links
}
\description{
A dataset containing the Jaccard coefficient for all pairwise comparisons
between all genesets. Self reflective edges and zero edges are removed. The
variables are as follows:
}
\keyword{datasets}

