% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.smi}
\alias{read.smi}
\title{Read a smiles file
returns data.frame with columns [smiles, compound]}
\usage{
read.smi(fname, smiles = "smiles", compound = "compound",
  fingerprint = NULL, data = NULL, check_duplicates = TRUE)
}
\description{
Read a smiles file
returns data.frame with columns [smiles, compound]
}
