% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{risk_set}
\alias{risk_set}
\alias{risk_set_size}
\title{Calculate Risk Set}
\usage{
risk_set(size, ..., sep = "|")

risk_set_size(size, ...)
}
\arguments{
\item{size}{a numeric vector}

\item{...}{vectors of identities}

\item{sep}{a string}
}
\value{
a risk set data frame

a vector of sizes
}
\description{
Calculate Risk Set

Risk Set Sizes
}
