% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaR-package.R
\docType{package}
\name{seaR-package}
\alias{seaR}
\alias{seaR-package}
\title{seaR: Structural Event Analysis}
\description{
Implements Martin Ruef's Structural Event Analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/acastroaraujo/seaR}
  \item Report bugs at \url{https://github.com/acastroaraujo/seaR/issues}
}

}
\author{
\strong{Maintainer}: Andrés Castro Araújo \email{acastroaraujo@gmail.com}

}
\keyword{internal}
